// src/interfaces/socket-events.interface.ts
export enum SocketEventName {
    eventsChange = 'eventsChange',
    faultsChange = 'faultsChange',
    alertsChange = 'alertsChange',
    toDeleteAlerts = 'toDeleteAlerts',
    systemsUnreadAlertsCount = 'systemsUnreadAlertsCount',
    userCommandsChange = 'userCommandsChange',
    commandResultTimeFilterChange = 'commandResultTimeFilterChange',
    startUpInBitSensors = 'startUpInBitSensors',
    graphInfoTimeFilterChange = 'graphInfoTimeFilterChange',
    treeChange = 'treeChange',
    serviceability = 'serviceability',
    clientVersion = 'clientVersion',
    serverHealth = 'serverHealth',
    mapsSelectionNames = 'mapsSelectionNames',
    systemStateChange = 'systemStateChange',
    commandOptions = 'commandOptions',
    sensorInBit = 'sensorInBit',
    systemInfo = 'systemInfo',
    keepAlive = 'keepAlive',
    communicationStatus = 'communicationStatus',
    graphInfo = 'graphInfo',
    treeMapOfflineServiceability = 'treeMapOfflineServiceability',
    bitReport = 'bitReport',
    commandsResults = 'commandsResults',
    systemLogTimeFilterChange = 'systemLogTimeFilterChange',
    treeMapFilterChange = 'treeMapFilterChange',
    serverConfigToClient = 'serverConfigToClient',
    clientsPeriodicalOnlineDataRemovelFromStore = 'clientsPeriodicalOnlineDataRemovelFromStore',
    updateNodeVisibility = 'updateNodeVisibility',
}